requirements.txt
Flask==3.0.3
flask-cors==3.0.9
numpy==2.2.4
open3d==0.19.0
pydicom==3.0.1
scikit-image==0.25.2
scipy==1.15.2
tqdm==4.67.1
numpy-stl==2.16.0
gunicorn==20.1.0